feat(api): Implement user order management and restaurant menu retrieval
- Added GET handler for fetching active orders for the authenticated user.
- Implemented GET handler for retrieving all orders for the authenticated user.
- Created GET handler for fetching restaurant menu items by restaurant ID.
- Developed GET handler for fetching public restaurant data including popular items.
- Introduced GET handler for fetching reviews from Google Places API.
- Added CRUD operations for user addresses with validation using Zod.
- Implemented like/unlike functionality for user menu items with appropriate handlers.
- Developed user profile retrieval and update functionality with validation.
- Created Stripe webhook handler for processing checkout session events and order creation.
