+++ next.config.ts
@@ -2,7 +2,7 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   images: {
-    domains: ['pixel.menusifu.com', "lh3.googleusercontent.com"],
+    domains: ["lh3.googleusercontent.com"],
     unoptimized: true,
   },
 };
